Implement Redis Foreign Data Wrapper functionality
- Add schema import functionality for Redis foreign tables, allowing users to import schemas based on Redis keys and types.
- Implement foreign truncate functionality to delete keys in Redis corresponding to foreign tables.
- Introduce comprehensive column validation tests for various Redis data types (string, hash, set, zset, list, stream) to ensure correct column configurations.
- Validate multi-key patterns and ensure proper handling of key prefixes in foreign table definitions.
