+++ src/core/mod.rs
@@ -1,12 +1,10 @@
-/// Core FDW functionality module
-///
-/// This module contains the essential components for the Redis Foreign Data Wrapper:
-/// - Connection factory for creating and configuring Redis connections
-/// - Global connection pool manager for efficient connection reuse
-/// - FDW handlers that integrate with PostgreSQL's foreign data wrapper infrastructure
-/// - State management for query execution
+pub mod column_utils;
 pub mod connection_factory;
+pub mod explain;
 pub mod handlers;
+pub mod join;
 pub mod pool_manager;
+pub mod schema_import;
 pub mod state_manager;
+pub mod truncate;
 pub mod validator;
